Updated currency stats to take doubles

+++ sdk-model/src/main/java/com/edukids/sdk/model/CurrencyStats.kt
@@ -5,7 +5,7 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class CurrencyStats(
-    val currentAmount: Long,
-    val earnedInInstance: Long,
-    val currentlyEarnable: Long
+    val currentAmount: Double,
+    val earnedInInstance: Double,
+    val currentlyEarnable: Double
 ) : Parcelable
